Swap eth for token with referral test

+++ src/interfaces/IConveyorSwapAggregator.sol
@@ -32,7 +32,7 @@ interface IConveyorSwapAggregator {
 
     function swapExactEthForTokenWithReferral(
         address tokenOut,
-        uint256 amountOutMin,
+        uint128 amountOutMin,
         uint128 protocolFee,
         ConveyorSwapAggregator.SwapAggregatorMulticall
             calldata swapAggregatorMulticall,
